Random Number Guessing Game
The computer randomly chooses a number with different ranges based on the difficulty the user selects. If the user selects the easy difficulty, the player will have 3 attempts to guess the computer's number, which is within the range of 1 - 10. If the user selects the medium difficulty, then the player will have 4 attempts and the computer's number will be within the range of 1 - 50. On the other hand, if the user selects the hard difficulty, then the player will have 5 attempts to guess within a range of 1 - 100. There are lines within the code that tells the user whether the computer's number is higher or lower than the user's guess. Messages are also displayed when the user either wins or loses in the game.

Basic Assignment Requirements:
1. Computer randomly generates a number from 1 - 10
2. User has one attempt to guess
3. Display "You, guessed correctly! You win!" if the player guesses correctly or "I'm sorry, that is not correct. You lose." if they choose incorrectly

The Extra Credit: 
1. Have several difficulties with the following conditions:
    - Easy: 3 tries, range 1 - 10
    - Medium: 4 tries, range from 1 - 50
    - Hard: 5 tries, range from 1 - 100
2. Tell the user whether the computer's number is higher or lower than the user's guess.