Rock, Paper, Scissors Game
This game implements switch statements that connect 1, 2, and 3 to the choices rock, paper, and scissors, respectively. By doing so, the user and computer may randomly choose a number from 1 to 3 to choose from rock, paper, or scissors. The program compares the two choices and prints who is the winner of the round. This code also tallies up all of the choices, wins, losses, and ties of both the player and the computer and displays them once the user chooses not to participate in the game anymore. Along with the wins, losses, and other statistics, the choice that the winner most commonly won with is displayed. If the computer was the winner overall, another message is displayed telling the player what choice that he or she should have chosen more often to win a greater amount of rounds than the computer. 

Basic Assignment Requirements:
1. I must use at least one switch statement and one if statement.
2. The player enters 1 for Rock, 2 for Paper, 3 for Scissors.
3. The computer randomly generates a number from 1 to 3.
4. The program displays who won the round or whether the round was a tie.
5. Ask the player if they would like to play again.
6. Keep count of the choices that the computer and the user had chosen and display them at the end of the round.
7. Keep count of the wins, losses, and ties of the user and computer and display them at the end of the round.
8. Use the rand() function to have the computer randomly select a number from 1 to 3.

The Extra Credit:
1. At the end of the game, display who won the most rounds and what choice the winner won the most rounds with.
2. Tell the player what they should have chosen to win most of the time. (For example, if the computer chose rock 7 out of the 10 times then the player should have chosen paper to win most of the time.)