Timed Multiplication Game
Uses do-while loops, if, else-if, and else statements, a timer is implemented. The switch statement allows the user to choose one of the three difficulties each with different factor ranges, allotted times, and score goals:
1 - Easy: 5 seconds, 5 correct, factors range from 1 to 5
2 - Medium: 3 seconds, 7 correct, factors range from 5 - 10
3 - Hard: 14 seconds, 20 correct, factors range from 10 - 15
Clock_gettime commands were used to get real wallclock time and beginning and end times were compared to create a timer.