Jurassic Park Text Adventure Game
This game uses pass by values, structures, bool values, int values, a string value, float values, switch statements, all if-else statements, as well as several functions to create a functioning segment of an adventure game that passes around a hero and its attributes. This segment gives the hero the necessary amount of money to successfully make it to the end and obtain the key. 

If the rest of the game was attached, the hero would have to run through the other three segments and obtain the remaining three keys to beat the entire adventure game (which has the premise of being transported into Netflix and going through 4 shows or movies to acquire keys to get out of Netflix). However, since this is my portion, I am only uploading this. Enjoy!